requirements.txt
Flask==2.0.2
pyserial==3.5
python-dotenv==0.19.2